fix: improve color scheme store perf
